movedstyle context and product state to App.jsx
