+++ client/src/components/Overview/productInfo/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Price from './price';
 import Share from './share';
 import Description from './description';
-import { styleContext } from '../index';
+import { styleContext } from '../../App';
 
 export default function ProductInfo({ product, avgRating }) {
   const { style } = React.useContext(styleContext);
